librados: Add opentelemtry trace param for aio_operate and operate methods
in order to propagate the trace info I added the otel-trace as an extra param. in some places, there already was a blkin trace info, and since it is not used in other places we can safely change it to o-tel trace info. but it will be done in another commit, so the cleanup of blkin trace will be in a dedicated commit

Signed-off-by: Omri Zeneva <ozeneva@redhat.com>