cmake: embed python dependencies mgr dashboard
With these changes python dependencies are embedded inside a
`__pypackages__` folder in pybind/mgr which will be preprended to the
`PYTHONPATH` so that that folder is the first one checked for
dependencies.

ceph.spec.in is updated to disable installing dashboard dependencies by
default and instead we provide a `--without embedded_pip_pkgs` option to disable embedded
dependencies.

New dependencies are added to pybind/mgr/requirements-mgr.txt and a lock
file is generated with pip-tools through a tox command: `tox -e gen-mgr-requirements`.

A script is introduced, in script/pypackages.sh, to generate the __pypackages__ folder which holds
the dependencies and later used by the mgr to import them.

Signed-off-by: Pere Diaz Bou <pdiazbou@redhat.com>