ceph-volume: system.get_mounts() refactor
When a network mount is present in `/proc/mounts` but for any reason
the corresponding server is down, this function hangs forever.
In a cluster deployed with cephadm, the consequence is that
it triggers `ceph-volume inventory` commands that hang and stay in D
state.

The idea here is to use a thread with a timeout to abort the call if the
timeout is reached.
`get_mounts()` is now a method of a class so we can exclude a path
altogether during the whole `inventory` execution (otherwise,
ceph-volume would try to access it as many devices there is on the
host which could slow down the inventory execution)

Resolves: rhbz#2152053

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>
(cherry picked from commit 94df47160a4292b2a554b7db9e27180cbf11b944)