rgw: fix improved admin token diagnostics.
Being picky about admin tokens has uncovered a further issue;
radosgw was quite willing to use a token that was on the cusp
of expiring.  Changed logic here will shorten expiration times
on tokens that radosgw can renew, which will cause it to renew
after either 5/6th of the lifetime (if the lifetime is less than 1 hour),
or 10 minutes before expiration (when lifetime >= 1 hour).

Signed-off-by: Marcus Watts <mwatts@redhat.com>
Resolves: rhbz#2099828
Resolves: rhbz#2123335