crimson/osd: vendor stop_signal.h
stop_signal.h is copied from seastar/apps/lib/stop_signal.hh. this
change copies it int our project so we can customize it in a following
commit. we will need to add an `abort_source` member to it.

the class `stop_signal` works great for us. but under some
circumstances, we also need to stop the crimson server programmatically
from the server itsef, for instance, per the request of monitor. in that
case, `stop_signal` cannot fulfill our needs, as the only thing which
can stop it is, SIGTERM or SIGTERM. we could send SIGINT to ourselves to
unblock `stop_signal::wait()`. but it would be better if we could leverage
`abort_source` for this purpose. for two reasons:

* `abort_source` allows use to cancel a "blocking" op.
  we have a couple background tasks in crimson, like `AdminSocket`'s
  accept and handle task, which could have been stopped by an
  `abort_source`. but now it checks for the `stop_gate` before accepting
  an incoming connection. this pattern cannot be always be repeated.
  because we cannot *abort* a "blocking" task. for instance, we use
  a homebrew `tri_mutex` to protect the read/write consistency of
  an obc, what if the server would need to stop when a request is
  still trying to acquire the lock? with the help of the `abort_source`,
  we should be able to either subscribe from the `abort_source`,
  and set an abort exception for each of the waiters. the same applies
  to other "blocking" calls like `shared_promise::get_shared_future()`,
  this allows us to abort the waiter before the promise is resolved.
  this could be handy if we need to abort potentially op which
  might take long and should be aborted when monitor wants to
  kill the osd instance.
* `abort_source` allows us to abort the services programmatically.
  this is the use case mentioned in the beginning of this commit
  message.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>