rbd: Moved get_image_or_snap_spec and parse_options to utils
At the moment, there are multiple copies of "get_image_or_snap_spec".

For this reason, we're moving it to Utils.cc so that all backends may
re-use it, also benefiting from the recent pool related fix.

We are also renaming parse_options to append_options_as_args and
changing its type to void

Signed-off-by: Stefan Chivu <schivu@cloudbasesolutions.com>