common: Optimize the HDD mClock cost parameters.
With the earlier cost per io and cost per byte parameters, the
client ops performance was good but the recovery ops were taking
much longer to complete due to the other optimizations (see
commit f4c4f0855f260e6daedd6d533eab4471b73164e7) made to
the recovery limit settings in the mclock profiles. The new cost
parameters optimizes this further so that the recovery ops
progress faster while maintaining similar client throughput.

This optimization is tested on both a small cluster and a larger
cluster in a scaled environment.

Fixes: https://tracker.ceph.com/issues/57529
Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>