install-deps: remove emacs modeline from script
The majority of the lines in this shell script uses spaces to indent,
but I *think* this modeline tells emacs to indent with tabs. In a very
near future change I'm going to make all indents use spaces. So we first
remove the emacs modeline that tells it to use the opposite.

Signed-off-by: John Mulligan <jmulligan@redhat.com>