mgr/cephadm: redo service level config when spec is updated
Previously, the service config function was only called
when we deploy a new daemon for that service. That meant
that updates to the spec such as changing a cert that don't
affect the daemon placement wouldn't trigger the service level
config to happen again. With this change, we now mark
the service as needing its config function ran if a daemon
for the service is added/removed or if the spec is updated.

Fixes: https://tracker.ceph.com/issues/58100

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 1d50a2cdfb32b60b352cc03a83f8d1c35f2ae178)