seastar: bump up seastar submodule to recent master
to address the fmtlib v9 compatibility issues.

please note, the submodule is bump up to a commit which includes
a revert commit on top of upstream master HEAD. the revert is to
work around the a recent change which adds a constraint on the parameter
type of handle_exception(). we are using errorator along with
handle_exception(), but errorator cannot fulfill the requirement of
seastar::Future(), so we need to drop the change before we have a
better solution.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>