mds: force replay sessionmap version
When expiring the MDLog Segments will it persist the sessionmap,
but what if the number of Segments doesn't reach up to the max
segment threshhold the expiring may won't happen.

So if the MDS failover then the sessionmap won't be persist and
couldn't be replayed by the standby MDS when it starting. We should
just force replay the sessionmap version just the inotable does.

Fixes: https://tracker.ceph.com/issues/58489
Signed-off-by: Xiubo Li <xiubli@redhat.com>