mgr/volumes: removing the helper from import which is not required
This commit fixes the bug arose due to the addition of a
patch by mistake that wasn't supposed to be the part of this.
i.e. the import of the helper method `set_group_attrs` which is
part of another code is mistakenly added in the MR pointing to
BZ#2122275.
Resolves: rhbz#2166652
Signed-off-by: Neeraj Pratap Singh <neesingh@redhat.com>