mgr/cephadm: add more aggressive force flag for host maintenance enter
Right now, we have safety checks that will either say things are okay to stop,
return warnings, or returns "alerts". Warnings can be bypassed already with
the --force flag that exists for the command. However, the alerts cannot be
bypassed and cephadm will refuse to attempt to put the host in maintenance mode.
The idea here is for an additional flag that also bypasses that alerts, in cases
where a user really needs to put the host into maintenance mode even if that
means causing issues within the cluster.

Fixes: https://tracker.ceph.com/issues/58892

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 0a980666f645387f1eaa5f2c995db304c42e85e6)