mgr/cephadm: minor reorganizing of nfs service test
This change makes some formatting and organizational changes to one of
the nfs + ingress test cases. This is in preparation for a future change
that will parametrize this function allowing it to test haproxy config
when proxy protocol is enabled and when not.

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 5ad7f2d0d57adb0a7f94403090ecb82c748c16c1)