rgw/multisite: parse shard_id value correctly to handle num_shards 0 case
For buckets that have num_shards set to 0, bucket instance
will not have a shard_id delimiter. When this bucket instance is parsed,
we end up assigning a value of -1 to shard_id, which is invalid
in data sync. This change ensures that we represent the shard_id correctly
by giving it a valid number

Signed-off-by: Shilpa Jagannath <smanjara@redhat.com>