crimson/osd/osd_operations/snaptrim_event: Handle missing clone obc
The case where resolve_oid is not able to return the clone oid (in order
to load the clone obc) should result in an SNAPTRIM_ERROR and stop the loop
iteration. Without taking this case into account an infinite loop of snap
trim events will be created.
A comment was also added to reference classic's approach for this case.

Fixes: https://tracker.ceph.com/issues/59231

Signed-off-by: Matan Breizman <mbreizma@redhat.com>