pybind/ceph_argparse: Error message is not descriptive for ceph tell command
$ceph tell cephfs_1.magna023.yvyxvl client ls
error handling command target: local variable 'poolid' referenced before assignment

the error message `local variable 'poolid' referenced before assignment` looks not  descriptive.
The exception was not getting raised by the usage of `ArgumentValid` and hence code asks for poolid which is
due to the bypass of exception when no type has been provided in tell commands.

Now after this fix, where we are raising JsonFormat Exception, it is not getting bypassed and hence
we are getting the proper error message when type is unknown.

Fixes: https://tracker.ceph.com/issues/59624
Signed-off-by: Neeraj Pratap Singh <neesingh@redhat.com>