mgr: block registration of a mgr module's  client
... until the registration is reflected in the new MgrMap received
from the monitor. This guarantees that the monitor can blocklist all
the registered clients of the mgr being dropped by the monitor.
And this also prevents the possible racing of the clients of the mgr
being failed that reconnect on being blocklisted against the clients
of the standby mgr being promoted during mgr failover.

When a module is registering a client, the client is first registered
in the mgr's py_module_registry. Next time the mgr sends a beacon to
the monitor, it requests the monitor to register the client in the
its py_module_registry onto the monitor's MgrMap. The monitor on
receiving the beacon registers the client in the MgrMap, and sends an
updated MgrMap to the mgr. It's possible that a mgr module registers
its client in the py_module_registy, and when waiting for the client
to be registered in the monitor's MgrMap another mgr thread deregisters
the client from the py_module_registry. Here, the mgr will never send
a beacon message requesting for the client to be registered in the
MgrMap as the client is no longer in the py_module_registry. So the
module will be stuck waiting for the MgrMap with the client registered
in it. To prevent such a scenario, error out of the module's client
registration after receiving the new MgrMap from the monitor and
noticing that the client is not in the mgr's py_module_registry.

Fixes: https://tracker.ceph.com/issues/58924
Signed-off-by: Ramana Raja <rraja@redhat.com>