mgr/cephadm: make jaege-collector a dep for jaeger-agent
the jaeger-agent's need to build a url for the collector(s)
that have been deployed. If a collector moves, or we deployed
the agents before the collector, we need to reconfig the agents
with updated info about the collectors. Failure to do so can
leave the jager-agents down reporting

```
Could not create collector proxy","error":"at least one collector hostPort address is required when resolver is not available"
```

Fixes: https://tracker.ceph.com/issues/59704

Signed-off-by: Adam King <adking@redhat.com>